test(e2e): migrate wallet smoke specs to Appium (MMQA-2006)
Move evm-provider-events and incoming-transactions from Detox to
smoke-appium, with Appium network-row selection and native-UI provider
assertions. Validated on iOS and Android.
